wip: minimize diffs
