+++ server/server_test.go
@@ -82,12 +82,9 @@ func RunServer(opts *Options) *Server {
 		panic(fmt.Sprintf("No NATS Server object returned: %v", err))
 	}
 
-	// if !opts.NoLog { <>/<>
-	opts.NoLog = false
-	opts.Trace = false
-	opts.Debug = false
-	s.ConfigureLogger()
-	// }
+	if !opts.NoLog {
+		s.ConfigureLogger()
+	}
 
 	// Run server in Go routine.
 	s.Start()
